Decided on logic Ordering
So I’ve decided that we’re going to implement the loading when the
navigation changes to the playlist ID page. This way people can
bookmark pages, and the loader will still show up, resulting in a
better result for the enduser. I don’t like the timeout right now, I’m
going to try to put an event listener instead, because this is going to
take 1s for playlists that might only be 1 item long.

Also setup the id for the playlists, I’m going to implement duplicate
elimination later to prevent the same playlist being loaded into local
storage twice.