All Song Data Collected and Persisted
So now the entire information gathering process is complete. All of the
songs are being got, the runtimes are in the correct order, and there
are no “Deleted” or “Private” videos in the playlist. An improvement on
previous iterations is that we can just completely ignore them as
opposed to including them in API calls and arrays.