LocalStorage Utilisation Partially Implemented
Now if the user has already had this playlist before, they can load the
songs directly from the session storage instead of pinging the API
which can take seconds to complete. One potential problem here is that
we don’t know if the playlist has been updated since the last time the
user used this, without pinging the server. I might implement a check
in the background to tell our program that the playlist has been
updated, and next time it should be gathered from scratch.