Uses LocalStorage, and checks for updates on the playlist
So the problem before is that the local storage essentially has a
snapshot of the playlist from when the user first issued the request,
but what if things change? Well now I have a check to see if the number
of songs change, and then I’ll set a bit on the playlist object to say
that it needs updating next time. The one issue now is if the user adds
and then removes a song, the total will appear the same. I’m thinking I
could do an MD5 hash on the playlist, and then compare the hashes of
the old and new playlists to see if a change is needed.