Progress on the player has begun
I’ve started trying to embed the youtube player in my site, I’m coming
across a few issues at the moment in incorporating the functions I need
as part of the view. I might just put the source code in the view file
to begin with and then get it to play nicely with backbone afterwards.